update fetch on game.tsx
